requirements.txt
# Core
anthropic>=0.39.0
python-dotenv>=1.0.0

# Stock data
yfinance>=0.2.31

# Technical analysis
pandas>=2.1.0
pandas-ta>=0.3.14b
numpy>=1.26.0

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
lxml>=4.9.0

# Thai NLP
pythainlp>=4.0.0
# transformers>=4.35.0  # for WangchanBERTa (uncomment when needed)
# torch>=2.1.0          # for WangchanBERTa (uncomment when needed)

# Alerts
requests>=2.31.0

# Scheduling
apscheduler>=3.10.0

# Config
pyyaml>=6.0.1

# Database
# sqlite3 is built-in

# Utilities
rich>=13.7.0
